Updated my code

+++ server/routes/manage.route.js
@@ -5,8 +5,9 @@ const CategoryModel = require('../models/category.model')
 
 router.post('/delete-task', async (req, res)=>{
     const task = req.body.task_name
+    const user = req.body.user_id
     try{
-        await TaskModel.deleteOne({task_name: task, done: 0})
+        await TaskModel.deleteOne({user_id: user, task_name: task, done: 0})
         .then(res.sendStatus(200))
     }
     catch{
@@ -15,9 +16,9 @@ router.post('/delete-task', async (req, res)=>{
 })
 
 router.post('/update-task', async (req, res)=>{
-    var {task_name, done} = req.body
+    var {user_id, task_name, done} = req.body
     try{
-        await TaskModel.updateOne({task_name: task_name, done: 0}, {done: done})
+        await TaskModel.updateOne({user_id: user_id, task_name: task_name, done: 0}, {done: done})
         .then(res.sendStatus(200))
     }
     catch{
@@ -28,8 +29,9 @@ router.post('/update-task', async (req, res)=>{
 
 router.post('/delete_cat', async (req, res)=>{
     const cat = req.body.name
+    const user = req.body.id
     try{
-        await CategoryModel.deleteOne({cat: cat})
+        await CategoryModel.deleteOne({cat: cat, user_id: user})
         .then(res.sendStatus(200))
     }
     catch{
